Split specs into separate files and refactor them
