+++ lib/google_contacts_api/group_set.rb
