+++ spec/fixtures/empty_contact_set.json
