+++ spec/fixtures/empty_group_set.json
@@ -1,0 +1,72 @@
+{
+    "version": "1.0",
+    "encoding": "UTF-8",
+    "feed": {
+        "xmlns": "http://www.w3.org/2005/Atom",
+        "xmlns$openSearch": "http://a9.com/-/spec/opensearch/1.1/",
+        "xmlns$gContact": "http://schemas.google.com/contact/2008",
+        "xmlns$batch": "http://schemas.google.com/gdata/batch",
+        "xmlns$gd": "http://schemas.google.com/g/2005",
+        "gd$etag": "W/\"C0UESXY5eip7ImA9WhdTEkw.\"",
+        "id": {
+            "$t": "example@gmail.com"
+        },
+        "updated": {
+            "$t": "2011-07-09T11:33:28.822Z"
+        },
+        "category": [{
+            "scheme": "http://schemas.google.com/g/2005#kind",
+            "term": "http://schemas.google.com/contact/2008#group"
+        }],
+        "title": {
+            "$t": "Example User's Contact Groups"
+        },
+        "link": [{
+            "rel": "alternate",
+            "type": "text/html",
+            "href": "http://www.google.com/"
+        },
+        {
+            "rel": "http://schemas.google.com/g/2005#feed",
+            "type": "application/atom+xml",
+            "href": "https://www.google.com/m8/feeds/groups/example%40gmail.com/full"
+        },
+        {
+            "rel": "http://schemas.google.com/g/2005#post",
+            "type": "application/atom+xml",
+            "href": "https://www.google.com/m8/feeds/groups/example%40gmail.com/full"
+        },
+        {
+            "rel": "http://schemas.google.com/g/2005#batch",
+            "type": "application/atom+xml",
+            "href": "https://www.google.com/m8/feeds/groups/example%40gmail.com/full/batch"
+        },
+        {
+            "rel": "self",
+            "type": "application/atom+xml",
+            "href": "https://www.google.com/m8/feeds/groups/example%40gmail.com/full?alt\u003djson\u0026max-results\u003d25"
+        }],
+        "author": [{
+            "name": {
+                "$t": "Example User"
+            },
+            "email": {
+                "$t": "example@gmail.com"
+            }
+        }],
+        "generator": {
+            "version": "1.0",
+            "uri": "http://www.google.com/m8/feeds",
+            "$t": "Contacts"
+        },
+        "openSearch$totalResults": {
+            "$t": "0"
+        },
+        "openSearch$startIndex": {
+            "$t": "1"
+        },
+        "openSearch$itemsPerPage": {
+            "$t": "25"
+        }
+    }
+}
