+++ spec/fixtures/group_set.json
