+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ require 'json'
 require 'hashie'
 require 'net/http'
 require 'google_contacts_api'
+require 'ostruct'
 Dir[File.dirname(__FILE__) + "/support/**/*.rb"].each {|f| require f }
 
 puts "Testing version #{GoogleContactsApi::Version::STRING}"
